Remove Analysis and Download, make it seperate feature requests in the future.
